# Documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Supchat
  description: une API de chat pour une entreprise
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Serveur local Symfony
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Users:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          minLength: 3
          maxLength: 255
        lastName:
          type: string
          minLength: 3
          maxLength: 255
        userName:
          type: string
          minLength: 3
          maxLength: 255
        emailAddress:
          type: string
          format: email
          minLength: 3
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 255
          nullable: true
        role:
          type: string
          enum: ["ROLE_USER", "ROLE_ADMIN"]
        theme:
          type: boolean
        status:
          type: string
          minLength: 3
          maxLength: 255
      required: [firstName, lastName, userName, emailAddress, role, theme, status]
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Paramètre de recherche manquant"
paths:
...
  /api/files/upload:
    post:
      tags: [Files]
      summary: Upload d’un fichier lié à un message
      operationId: uploadFile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, message_id]
              properties:
                file:
                  type: string
                  format: binary
                message_id:
                  type: integer
      responses:
        '201':
          description: Fichier uploadé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  filePath:
                    type: string
        '400':
          description: Fichier ou message_id manquant ou type invalide
        '404':
          description: Message introuvable
        '500':
          description: Échec de l’upload
  /api/files/me:
    get:
      tags: [Files]
      summary: Liste des fichiers que j’ai envoyés
      operationId: getMyFiles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des fichiers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    path:
                      type: string
                    name:
                      type: string
  /api/files/shared:
    get:
      tags: [Files]
      summary: Fichiers envoyés par d'autres dans mes workspaces
      operationId: getSharedFiles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des fichiers partagés
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    path:
                      type: string
                    name:
                      type: string
                    author:
                      type: string
  /api/files/{id}/generate-download-url:
    get:
      tags: [Files]
      summary: Génère une URL temporaire de téléchargement sécurisé
      operationId: generateDownloadUrl
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: URL générée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '403':
          description: Signature invalide ou lien expiré
        '404':
          description: Fichier introuvable
  /api/files/download:
    get:
      tags: [Files]
      summary: Téléchargement sécurisé via URL signée
      operationId: downloadFile
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
        - name: user
          in: query
          required: true
          schema:
            type: integer
        - name: ts
          in: query
          required: true
          schema:
            type: integer
        - name: _hash
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fichier téléchargé
        '400':
          description: Paramètres manquants
        '403':
          description: Signature invalide ou expirée
        '404':
          description: Fichier non trouvé
  /api/files/{id}:
    delete:
      tags: [Files]
      summary: Supprime un fichier
      operationId: deleteFile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Fichier supprimé
        '403':
          description: Accès interdit
        '404':
          description: Fichier introuvable
...
  /api/messages/{id}:
    get:
      tags: [Messages]
      summary: Récupère un message
      operationId: getMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Message trouvé
        '404':
          description: Message introuvable

    put:
      tags: [Messages]
      summary: Met à jour le contenu d’un message
      operationId: updateMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Message modifié
        '400':
          description: Contenu invalide

    delete:
      tags: [Messages]
      summary: Supprime un message
      operationId: deleteMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Message supprimé
...
  /api/user:
    get:
      tags: [Users]
      summary: Infos de l'utilisateur connecté
      operationId: getUserInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Infos de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '401':
          description: Non connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Users]
      summary: Modifier les infos de l'utilisateur connecté
      operationId: updateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '200':
          description: Mise à jour réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Non connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Users]
      summary: Supprimer son propre compte
      operationId: deleteUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Compte supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Non connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...